Improve session binding accuracy
The TPM tracks which entity is bound to a session using a combination of the entity's name and it's auth value
at session creation. Mirror this behaviour on the host side, which allows us to drop a hack in
HierarchyChangeAuth where we had to clear the bound entity from the session context after changing the
hierarchy auth so that the session would continue to work (but only if the auth value isn't changed to the value
it was at session creation).