Fix some bugs when verifying response HMAC during authorizaion value changes
For functions that change an authoization value, we copy the Session struct to incorporate the new auth value
before verifying the response HMACs. However, we weren't preserving the value of the includeAuthValue property
when doing this. In scenarios where includeAuthValue is set during the command execution (ie, for unbound
sessions), and the command changes the auth value of the authorized entity from one value to another that isn't
the empty buffer, we don't compute the correct session key in processResponseSessionAuth because the incorrect
value of includeAuthValue means we don't include the new authorization value.

However, due to another bug in processResponseSessionAuth, we were bailing out when computing an empty session
key, which is always the case for unbound and unsalted sessions where Session.includeAuthValue is false.

Unfortunately, this wasn't caught by any tests - it requires an unbound and salted session (so there is always
a session key) and to change the auth value for the authorized entity to a non-empty value.