Drop nil pointer checks on function parameters where unnecessary
The marshalling code will already marshal an element's zero value when it encounters a nil pointer to that
element. If the nil pointer is wrapped in another struct via the "sized" tag (as *Public, *NVPublic and others
are), then the marshalling code handles this by marshalling a zero sized buffer.

For the cases where a struct has a union field, and the zero value for the corresponding selector field is an
invalid value for the union, this approach triggers errors in the marshalling code about invalid selector values.
These errors contain a lot of context if they are nested and can be a bit confusing. To avoid this, we just
ignore selector value errors during marshalling (but not during unmarshalling) and marshal an empty value (ie,
nothing).

The result of this is that nil pointers for parameters that aren't optional will be picked up by the TPMs
unmarshalling and parameter checking, and the TPM will return an error where mandatory parameters are missing.