Provide more useful errors
The previous UnmarshallingError type wasn't particularly useful because you
couldn't determine whether unmarshalling failed because response parameters
passed from the caller were bad, or because of an actual broken response.
This removes UnmarshallingError and introduces InvalidResponseHeaderError
and InvalidResponsePayloadError. If an error is likely to be cause by a
broken response, then one of these errors is returned, otherwise we just
return an error with fmt.Errorf.