Clean up error code types and improve error documentation
The TPM can return some format 1 errors that don't correspond to a handle, session or parameter,
so replace the separate ErrorCode0 and ErrorCode1 types with a single ErrorCode type and allow
TPMError to be used for both format 0 errors, and format 1 errors that aren't associated with a
handle, session or parameter.

Also improve the documentation for errors.