Make it possible to flush a saved session without context loading it first
TPMContext.ContextSave invalidated the ResourceContext if it corresponded to a session,
and it was not possible to create a ResourceContext for a session with TPMContext.WrapHandle.
This meant that that the only way to flush a saved session was to context load it with the
saved context using TPMContext.ContextLoad (which would create a new ResourceContext that
could be used in TPMContext.FlushContext).

This commit adds some extra state to the ResourceContext implementation for session contexts that
indicates whether the session is loaded and if the context contains all of the host-side state
required to use the session. When a session's context is saved, mark the ResourceContext as
unloaded rather than invalidating it.

When a session's context is loaded, determine if there is a (saved) ResourceContext associated
with the saved session context and use that instead of creating a new one.

Make it possible to create ResourceContext instances for sessions with TPMContext.WrapHandle,
by returning a ResourceContext that doesn't contain all of the host-side state required to use
the session. These ResourceContext instances can either be used to flush the session from the
TPM, or can be upgraded to a full ResourceContext if it corresponds to a saved session and the
saved session's context is loaded again.

This also changes how session ResourceContext instances are indexed by the tracking TPMContext -
session handles are always normalized to a HMAC session handle type, as this simplifies the fact
that ResourceContext instances associated with saved sessions that weren't created by this API
could correspond to HMAC or policy sessions.