Improve parameter encryption tests
When a dedicated encryption session is used, a nonce value related to the encryption
session is included in the HMAC computation for the first authorization session. There
was previously no test to make sure that this works correctly because the tests with a
dedicated encryption session only used a password authorization.