Simplify TPMContext.Startup
This had a bunch of complex logic to keep the TPMContext state in sync, but this is
crazy because on a real device, Startup will only be called after a reset and before
any other commands executed. Just evict all ResourceContext instances from the
TPMContext instead, to handle the simulator case which can be "power-cycled" without
creating a new TPMContext.