Various changes to ResourceContext management
This makes a few changes to how ResourceContext management works:
- If the TPM creates a resource with a handle that we already have a ResourceContext for, then we
just consider the old one stale and invalidate it rather than calling panic(). It's possible for
this to happen when working with sessions and persistent resources via a resource manager, as
other clients can flush / evict persistent resources that we have ResourceContexts for.
- There is some additional input sanitization in TPMContext.ContextLoad.
- Functions that create ResourceContexts (TPMContext.Load/LoadExternal/CreatePrimary/CreateLoaded/
ContextLoad) now check that the handle returned from the TPM is consistent with the type of resource.
- TPMContext.CreatePrimary/CreateLoaded make sure that the returned name is consistent with the
returned public area.
- TPMContext.Load/LoadExternal make sure that the returned name is consistent with the provided public
area.