Drop all of the complicated tracking of HandleContexts from TPMContext
TPMContext kept track of HandleContext instances created by it, enforced a
1-context-per-handle mapping and tried to keep tracked HandleContexts in sync
with their corresponding TPM entities when executing commands that have effects
on entities other than those specified in the command. This is unnecessary for
a couple of reasons:

- HandleContexts can be created from deserialized data rather than from a
public area read from the TPM, so there's already no guarantee these actually
correspond to resource on the TPM.
- HandleContexts can already be out of sync when connecting to the TPM through
a resource manager, where those contexts correspond to persistent objects, NV
indices or sessions.

In addition to this, the overall design is slightly problematic and slightly
weird. TPMContext.GetOrCreateResourceContext() would return a new
ResourceContext if one didn't already exist for the specified handle, else it
would return the existing one. But did it create a new one or not? If not, how
and when was it created? Was it created by reading the public area back from
the TPM? Was it created from a previously serialized ResourceContext? Was it
read back from the TPM with a session? These issues are more problematic when
passing TPMContext across package API boundaries - did another package create
the ResourceContext for a particular handle, and how? Also, if package A
passes TPMContext to package B, package B then uses
TPMContext.GetOrCreateResourceContext() to return a ResourceContext that was
created by package A, and then calls ResourceContext.SetAuthValue() in order
to use it - is this change being leaked to package B expected behaviour or
not?

This change drops all of the tracking of HandleContexts from TPMContext.
HandleContexts can only be created now. It also modifies TPMContext.Startup(),
TPMContext.Clear() and TPMContext.NVGlobalWriteLock() to not make changes to
HandleContexts corresponding to entities that are changed as a result of
those commands.