Make it easier to determine the maximum sizes of various types
This adds a EventMaxSize constant, indicating the maximum size of arguments
of the Event type. It also adds various helpers around TPM2_GetCapability
for detering the maximum sizes of various types.

Also added is TPMContext.InitProperties, currently used to initialize the
maximum NV buffer size. This allows initialization of properties on
TPMContext to be audited, and to preserve the exclusivity of an audit
session. The function is called automatically without any sessions when
the properties are required.