testutil: Implement some resource management
This improves the proxy transmission interface to implement
resource management to ensure that the TPM state is restored
where possible at the end of the test, and cleans up the
permissions a bit, which are now used to either request access
to any of the TPMs hierarchies or to request the ability to make
changes to the TPM's state that cannot be undone automatically by
the test harness.

This will improve the experience when running tests against a
real TPM device - something which isn't currently done, and which
means that transmission interface that is normally used is not
actually properly tested.