mu: Refactor how context is added to errors
The error handling currently adds context to errors using xerrors as
the stack unwinds, but this results in long error messages that are
very verbose, with the most important part (the actual error) being
right at the end.

This changes it to add all of the context at the point at which the
error occurs, and to just return this error without further wrapping.

This allows error messages to start with the most important part
(the error), and allows formatting of a parsing stack in the message
for more context.