Use HashAlgorithmId instead of crypto.Hash in various places
Some recent commits changed HashAlgorithmId to crypto.Hash in
some public APIs (mostly in util/), but this isn't really the
right thing to do - it creates inconsistency (some arguments
are represented by go-tpm2 types and others by go types),
and the TPM spec currently has one hash algorithm with no
corresponding crypto.Hash value (SM3). Whilst it's not possible
to use this algorithm in go-tpm2 right now even though there
is a go implementation of it (it requires it to be registered
with the crypto package), it may be possible to support it
in the future regardless of whether it gets a corresponding
crypto.Hash value.

Switch the use of crypto.Hash back to HashAlgorithmId in
various places.