mu: Special case sized and raw types in DetermineTPMKind
DetermineTPMKind had code to unwrap single field structures and
return the type of the single field in order to handle types that
are made sized or raw by wrapping them in a single field structure
with the appropriate tag. This predated the Sized() and Raw() APIs
though, and the behaviour is a bit weird for the non-sized and
non-raw cases - ie, if you pass a structure, it should return
TPMKindStruct whether it has one field or not.

Special case types that are wrapped with Sized() or Raw() instead.