mu: remove support for empty interface types
This only existed to handle the Raw() and Sized() APIs, which returned
a structure with an empty interface member and the appropriate tags set.
Interfaces generally don't make sense because we have no way to determine
what the concrete type should be during unmarshalling.

Instead, have Raw() and Sized() return a different opaque type that wraps
the real value and defines the appropriate options, and then special case
this before calling marshalValue / unmarshalValue.