mu: add IsSupported
Making DetermineTPMKind recurse in order to determine whether values
of the supplied type was a bit weird, because it could return
TPMKindUnion and values of this type will result in a panic if they
are used directly.

Revert DetermineTPMKind back to the old behaviour of only checking
the type of the supplied value without recursing, and introduce
IsSupported instead.