Don't error when unserializing a saved session context with CreateHandleContextFrom{Bytes,Reader}
This fails because a session context has the pointer to its session data
cleared during a TPM2_ContextSave, but the nil pointer is marshalled and
unmarshalled to a pointer to the zero value for the context data, which
fails unmarshalling due to the invalid digest algorithm (zero not being
valid).

It's trivial to fix this by making the session context data a sized type,
although there isn't really a lot of point in serializing and unserializing
a saved SessionContext.