Stop using MustMarshalTo* in a few places
The MustMarshalTo* functions should only be used when the parameters can
be guaranteed to marshal successfully (either because they are types
that can't fail or because we trust the data source), the API is
documented to panic on invalid inputs, or in test code.

There's a few uses where this isn't the case, so change them.