Make Public and NVPublic implement util.Entity
This makes Public and NVPublic implement the util.Entity interface,
which represents entities that have names, and means that these types
can be used along with HandleContext and Name in APIs that accept this
interface.

Unfortunately, the name is computed from these types which can fail.
The implementation of util.Entity in this case just returns Name(nil),
which is an invalid name and can be detected as such.

Unfortunately, there still needs to be a mechanism to compute the name
that can return an error. The original Name() functions have been
renamed to ComputeName().

This is an unfortunate API break, but should hopefully be the last one
before I tag a stable 1.0 release :/