Support setting command timeouts
This adds a new API - TPMContext.SetCommandTimeout that allows an
application to specify a timeout. The current behaviour is that reads
can block indefinitely, and that is still the default.

This also fixes a minor issue in the linux backend - we currently access
the raw file descriptor for polling because go's netpoller tries to read
before parking the current go routine, and the Linux TPM character
device does not return -EWOULDBLOCK if there is no data to be read yet
which results in the read always returning io.EOF on the go side.
Accessing the raw file descriptor marks it as blocking, which means
subsequent commands block in the write rather than performing a
non-blocking write and then blocking before the read.