Rename handleContext.checkConsistency to checkValid and simplify
The only things this needs to check really are that enums are valid
values and that associated digest algorithms are available in order to
avoid panics when the context is used.

Checking that the name is consistent isn't really necessary, especially
when NewObjectResourceContext and NewNVIndexResourceContext can be used
to create contexts that could be serialized and then fail this check
during subsequent unserialization.

Other checks on the lengths of the session key and session nonce aren't
necessary - it doesn't confirm that they are actually valid. Checking
that the symmetric algorithm is a valid value isn't necessary either -
if it's not, then it fails unserialization.