Add new policyutil submodule
The policyutil submodule is for computing and working with authorization
policies, and will replace the functionality in the existing util
submodule.

The new API for computing authorization policies also creates
a serializable policy structure that can be loaded and executed later on
rather than having to call policy assertions manually.

There are some missing features right now - there is no branch support
or support for TPM2_PolicyAuthorize.