policyutil: remove PolicySecretParms
This provided a way to set the cpHash and expiration, as these aren't
bound to TPM2_PolicySecret and so they don't make sense to keep inside
the policy. However, it's just an additional piece of complexity and
doesn't provide a way for an authority to require these like
TPM2_PolicySigned does. This binding could technically done by using an
authorized policy which contained a TPM2_PolicyCpHash assertion for the
secret resource, but we don't support using policies for those at the
moment. When we do, add cpHash to the policy for TPM2_PolicySecret.