Remove the configurable timeout
It doesn't really make sense having this as long as there is no
mechanism to resume a command that timed out. It may make more sense in
the future to have a command submission timeout, which times out if a
command is not actually submitted to the TPM within a certain timeframe -
useful where access to the TPM goes through a resource manager. But this
would require support from the backend.