policyutil: refactor policy building API a bit
This renames PolicyBuilder.Policy to PolicyBuilder.Build, and merges
the steps of creating a policy and computing the digest for it, so
PolicyBuilder.Build is sufficient to create a fully computed policy.

Policy.Compute is now renamed to Policy.AddDigest to better reflect
what the API does.

The parameters for TPM2_PolicyCpHash and TPM2_PolicyNameHash have been
removed from the policy elements. These are stored on the PolicyBuilder.
This means that Policy.AddDigest doesn't work for policies that contain
these assertions, but it doesn't make sense to add a digest for these
policies anyway, as they are bound to a specific resource already.

As the unit tests need to make direct use of TPM2_PolicyOR which relied
on a test-only export, there is now a proper API for doing this
(NewPolicyBuilderOR).