Add extra private type field to HandleContext
There used to be something similar, but then it was removed and the
handle type was used to select the union contents, but then these
fields needed to be sized because they may or may not be empty depending
on the context. Instead, reintroduce a distinct context type that is
separate from the handle type, have this correspond to concrete go types,
and have some of these types map to nothing in the union. This way, we
can guarantee that ObjectContext.Public will always return something,
except after HandleContext.Dispose is called, and we can avoid returning
ObjectContext and NVIndexContext implementations where it really doesn't
make sense.