testutil: Deprecate the TCTI field of TPMTest in favour of Transport
The term "TCTI" was copied from TSS but doesn't make sense here, where
we are using the term "Transport" instead for the channel that
communicates with the TPM device.

Note that this does add a bit of complexity because the test suite
needs to keep both fields in sync for backwards compatibility.