internal/transportutil: implement a way to multiplex access to a transport
The mssim passes the TPM connection to the retrier, which accesses the
connection on a dedicated goroutine. However, the existing mssim
Transport API provides a way to send control commands via the TPM
connection (currently only the stop command, but lkely more in the
future). This introduces a race condition where the TPM connection is
used from 2 goroutines.

This introduces another API (NewMultiplexedTransportManager and
LockableTransport) which provides a way for multiple goroutines to be
able to access a single transport, using a LockableTransport for each
goroutine and access serialized with a locking mechanism.