Don't always assume that we can defer to TPM_RH_NULL when passed a nil ResourceContext
In general, the methods of TPMContext will treat a nil ResourceContext
as being the ResourceContext for TPM_RH_NULL.

This isn't always safe though - ie, for any commands that are expecting
a hierarchy (TPMI_RH_HIERARCHY), in order to prevent silent and
unintentional use of the null hierarchy, a nil ResourceContext for an
argument that corresponds to a hierarchy will be treated as the
ResourceContext for TPM_RH_UNASSIGNED, which will result in the TPM
responding with an error.

This only changes the behaviour of TPMContext.CreatePrimary. The same
will need to be done for TPMContext.CreateLoaded if we support it in the
future.