testutil: duplicateTransport should block read/write when closed
The duplicateTransport internal type, obtained by opening a
TransportBackedDevice, has a concept of being closed even if the
underlying transport isn't closed. In this case, the transport should
block access to read / write. This makes it more likely for tests to
catch cases where operations are performed on an inadvertently closed
transport.